requirements.txt
eitaa
requests
python-dotenv
groq
redis